do some code reorganization
